+++ py/desitarget/targetmask.py
@@ -8,31 +8,32 @@ from desiutil.bitmask import BitMask
 import yaml
 from pkg_resources import resource_filename
 
-
-_bitdefs = None
-
-
-def _load_bits():
+def load_mask_bits(prefix=""):
     """Load bit definitions from yaml file.
     """
-    global _bitdefs
-    if _bitdefs is None:
-        _filepath = resource_filename('desitarget', "data/targetmask.yaml")
-        with open(_filepath) as fx:
-            _bitdefs = yaml.load(fx)
+    us = ""
+    if len(prefix) > 0:
+        us = '_'
+    prename = prefix+us
+    fn = "{}/data/{}targetmask.yaml".format(prefix,prename)
+    _filepath = resource_filename('desitarget', fn)
+    with open(_filepath) as fx:
+        bitdefs = yaml.load(fx)
         try:
-            _load_priorities()
+            bitdefs = _load_mask_priorities(bitdefs,handle="priorities",prename=prename)
         except TypeError:
             pass
-    return
+        try:
+            bitdefs = _load_mask_priorities(bitdefs,handle="numobs",prename=prename)
+        except TypeError:
+            pass
+    return bitdefs
 
-
-def _load_priorities():
-    """Priorities are defined in the yaml file, but they aren't a bitmask
-    and they require some extra processing.
+def _load_mask_priorities(bitdefs,handle="priorities",prename=""):
+    """Priorities and NUMOBS are defined in the yaml file, but they aren't 
+    a bitmask and so require some extra processing.
     """
-    global _bitdefs
-    for maskname, priorities in _bitdefs['priorities'].items():
+    for maskname, priorities in bitdefs[handle].items():
         for bitname in priorities:
             #- "SAME_AS_XXX" enables one bit to inherit priorities from another
             if isinstance(priorities[bitname], str) and \
@@ -41,28 +42,31 @@ def _load_priorities():
                     priorities[bitname] = priorities[other]
 
             #- fill in default "more" priority to be same as "unobs"
-            if isinstance(priorities[bitname], dict):
-                if 'MORE_ZWARN' not in priorities[bitname]:
-                    priorities[bitname]['MORE_ZWARN'] = priorities[bitname]['UNOBS']
-                if 'MORE_ZGOOD' not in priorities[bitname]:
-                    priorities[bitname]['MORE_ZGOOD'] = priorities[bitname]['UNOBS']
+            #ADM specifically applies to dictionary of priorities
+            if handle=='priorities':
+                if isinstance(priorities[bitname], dict):
+                    if 'MORE_ZWARN' not in priorities[bitname]:
+                        priorities[bitname]['MORE_ZWARN'] = priorities[bitname]['UNOBS']
+                    if 'MORE_ZGOOD' not in priorities[bitname]:
+                        priorities[bitname]['MORE_ZGOOD'] = priorities[bitname]['UNOBS']
 
-                #- fill in other states as priority=1
-                for state, blat, foo in _bitdefs['obsmask']:
-                    if state not in priorities[bitname]:
-                        priorities[bitname][state] = 1
-            else:
-                priorities[bitname] = dict()
+                    #- fill in other states as priority=1
+                    for state, blat, foo in bitdefs[prename+'obsmask']:
+                        if state not in priorities[bitname]:
+                            priorities[bitname][state] = 1
+                else:
+                    priorities[bitname] = dict()
 
         #- add to the extra info dictionary for this target mask
-        for bitdef in _bitdefs[maskname]:
+        for bitdef in bitdefs[maskname]:
             bitname = bitdef[0]
-            bitdef[3]['priorities'] = priorities[bitname]
-    return
+            bitdef[3][handle] = priorities[bitname]
+    return bitdefs
 
 #- convert to BitMask objects
-if _bitdefs is None:
-    _load_bits()
+#if bitdefs is None:
+#    load_bits()
+_bitdefs = load_mask_bits()
 try:
     desi_mask = BitMask('desi_mask', _bitdefs)
     mws_mask = BitMask('mws_mask', _bitdefs)