+++ py/desitarget/test/test_priorities.py
@@ -104,10 +104,9 @@ class TestPriorities(unittest.TestCase):
     def test_mask_priorities(self):
         for mask in [desi_mask, bgs_mask, mws_mask]:
             for name in mask.names():
-                if name == 'SKY' or name == 'NO_TARGET' or name.startswith('STD') \
-                    or name in ['BGS_ANY', 'MWS_ANY', 'SECONDARY_ANY',
+                if name.startswith('STD') or name in ['BGS_ANY', 'MWS_ANY', 'SECONDARY_ANY',
                                 'IN_BRIGHT_OBJECT', 'NEAR_BRIGHT_OBJECT',
-                                'BRIGHT_OBJECT']:
+                                'BRIGHT_OBJECT', 'SKY', 'SV', 'NO_TARGET']:
                     self.assertEqual(mask[name].priorities, {}, 'mask.{} has priorities?'.format(name))
                 else:
                     for state in obsmask.names():